Refactor factories.
