+++ sequence/entanglement_management/generation/BarretKok.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
-from enum import Enum, auto
-from math import sqrt
+
 from typing import TYPE_CHECKING, Any
 
 from .generation_message import EntanglementGenerationMessage, GenerationMsgType, valid_trigger_time
@@ -10,14 +9,15 @@ if TYPE_CHECKING:
     from ...topology.node import Node
 
 from ...resource_management.memory_manager import MemoryInfo
-from ..entanglement_protocol import EntanglementProtocol
+
+from .generation_a import EntanglementGenerationA
+
 from ...kernel.event import Event
 from ...kernel.process import Process
-from ...components.circuit import Circuit
 from ...utils import log
 
 
-class EntanglementGenerationBarretKokA(EntanglementProtocol):
+class BarretKokA(EntanglementGenerationA):
     """Entanglement generation protocol for quantum router.
 
     The EntanglementGenerationA protocol should be instantiated on a quantum router node.
@@ -31,13 +31,7 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
         memory (Memory): quantum memory object to attempt entanglement for.
     """
 
-    _plus_state = [sqrt(1/2), sqrt(1/2)]
-    _flip_circuit = Circuit(1)
-    _flip_circuit.x(0)
-    _z_circuit = Circuit(1)
-    _z_circuit.z(0)
-
-    def __init__(self, owner: "Node", name: str, middle: str, other: str, memory: "Memory"):
+    def __init__(self, owner: "Node", name: str, middle: str, other: str, memory: "Memory", **kwargs: Any):
         """Constructor for entanglement generation A class.
 
         Args:
@@ -48,72 +42,11 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
             memory (Memory): memory to entangle.
         """
 
-        super().__init__(owner, name)
-        self.middle: str = middle
-        self.remote_node_name: str = other
-        self.remote_protocol_name: str = ''
+        # raise error if passed kwargs
+        if kwargs:
+            raise ValueError(f"Unexpected keyword arguments: {kwargs}")
 
-        # memory info
-        self.memory: Memory = memory
-        self.memories: list[Memory] = [memory]
-        self.remote_memo_id: str = ""  # memory index used by corresponding protocol on other node
-
-        # network and hardware info
-        self.fidelity: float = memory.raw_fidelity
-        self.qc_delay: int = 0
-        self.expected_time: int = -1   # expected time for middle BSM node to receive the photon
-
-        # memory internal info
-        self.ent_round = 0  # keep track of current stage of protocol
-        self.bsm_res = [-1, -1]  # keep track of bsm measurements to distinguish Psi+ and Psi-
-
-        self.scheduled_events = []
-
-        # misc
-        self.primary: bool = False  # one end node is the "primary" that initiates negotiation
-        self.debug: bool = False
-
-        self._qstate_key: int = self.memory.qstate_key
-
-    def set_others(self, protocol: str, node: str, memories: list[str]) -> None:
-        """Method to set other entanglement protocol instance.
-
-        Args:
-            protocol (str): other protocol name.
-            node (str): other node name.
-            memories (list[str]): the list of memory names used on other node.
-        """
-        assert self.remote_protocol_name == ''
-        self.remote_protocol_name = protocol
-        self.remote_memo_id = memories[0]
-        self.primary = self.owner.name > self.remote_node_name
-
-    def start(self) -> None:
-        """Method to start "one round" in the entanglement generation protocol (there are two rounds in Barrett-Kok).
-
-        Will start negotiations with other protocol (if primary).
-
-        Side Effects:
-            Will send message through attached node.
-        """
-
-        log.logger.info(f"{self.name} protocol start with partner {self.remote_protocol_name}")
-
-        # to avoid start after remove protocol
-        if self not in self.owner.protocols:
-            return
-
-        # update memory, and if necessary start negotiations for round
-        if self.update_memory() and self.primary:
-            # send NEGOTIATE message
-            self.qc_delay = self.owner.qchannels[self.middle].delay
-            frequency = self.memory.frequency
-            message = EntanglementGenerationMessage(GenerationMsgType.NEGOTIATE,
-                                                    self.remote_protocol_name,
-                                                    protocol_type=self,
-                                                    qc_delay=self.qc_delay,
-                                                    frequency=frequency)
-            self.owner.send_message(self.remote_node_name, message)
+        super().__init__(owner, name, middle, other, memory)
 
     def update_memory(self) -> bool | None:
         """Method to handle necessary memory operations.
@@ -156,7 +89,6 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
             self._entanglement_fail()
             return False
 
-
     def emit_event(self) -> None:
         """Method to set up memory and emit photons.
 
@@ -193,7 +125,7 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
         msg_type = msg.msg_type
 
         log.logger.debug("{} {} received message from node {} of type {}, round={}".format(
-                         self.owner.name, self.name, src, msg.msg_type, self.ent_round))
+            self.owner.name, self.name, src, msg.msg_type, self.ent_round))
 
         if msg_type is GenerationMsgType.NEGOTIATE:  # primary -> non-primary
             # configure params
@@ -203,7 +135,8 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
 
             # get time for first excite event
             memory_excite_time = self.memory.next_excite_time
-            min_time = max(self.owner.timeline.now(), memory_excite_time) + other_qc_delay - self.qc_delay + cc_delay  # cc_delay time for NEGOTIATE_ACK
+            min_time = max(self.owner.timeline.now(),
+                           memory_excite_time) + other_qc_delay - self.qc_delay + cc_delay  # cc_delay time for NEGOTIATE_ACK
             emit_time = self.owner.schedule_qubit(self.middle, min_time)  # used to send memory
             self.expected_time = emit_time + self.qc_delay  # expected time for middle BSM node to receive the photon
 
@@ -223,7 +156,8 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
 
             # schedule start if necessary (current is first round, need second round), else schedule update_memory (currently second round)
             # TODO: base future start time on resolution
-            future_start_time = self.expected_time + self.owner.cchannels[self.middle].delay + 10  # delay is for sending the BSM_RES to end nodes, 10 is a small gap
+            future_start_time = self.expected_time + self.owner.cchannels[
+                self.middle].delay + 10  # delay is for sending the BSM_RES to end nodes, 10 is a small gap
             if self.ent_round == 1:
                 process = Process(self, "start", [])  # for the second round
             else:
@@ -268,7 +202,7 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
             resolution = msg.resolution
 
             log.logger.debug("{} received MEAS_RES={} at time={:,}, expected={:,}, resolution={}, round={}".format(
-                             self.owner.name, detector, time, self.expected_time, resolution, self.ent_round))
+                self.owner.name, detector, time, self.expected_time, resolution, self.ent_round))
 
             if valid_trigger_time(time, self.expected_time, resolution):
                 # record result if we don't already have one
@@ -283,31 +217,10 @@ class EntanglementGenerationBarretKokA(EntanglementProtocol):
         else:
             raise Exception("Invalid message {} received by EG on node {}".format(msg_type, self.owner.name))
 
-    def is_ready(self) -> bool:
-        return self.remote_protocol_name != ''
-
-    def memory_expire(self, memory: "Memory") -> None:
-        """Method to receive expired memories."""
-
-        assert memory == self.memory
-
-        self.update_resource_manager(memory, MemoryInfo.RAW)
-        for event in self.scheduled_events:
-            if event.time >= self.owner.timeline.now():
-                self.owner.timeline.remove_event(event)
-
     def _entanglement_succeed(self):
-        log.logger.info(self.owner.name + " successful entanglement of memory {}".format(self.memory))
-        self.memory.entangled_memory["node_id"] = self.remote_node_name
-        self.memory.entangled_memory["memo_id"] = self.remote_memo_id
+        super()._entanglement_succeed()
         self.memory.fidelity = self.memory.raw_fidelity
-
         self.update_resource_manager(self.memory, MemoryInfo.ENTANGLED)
 
-    def _entanglement_fail(self):
-        for event in self.scheduled_events:
-            self.owner.timeline.remove_event(event)
-        log.logger.info(self.owner.name + " failed entanglement of memory {}".format(self.memory))
-
-        self.update_resource_manager(self.memory, MemoryInfo.RAW)
+EntanglementGenerationA.register('barretkokA', BarretKokA)
